requirements.txt
aiohttp==3.8.6
aiosignal==1.3.1
annotated-types==0.6.0
anyio==3.7.1
async-timeout==4.0.3
attrs==23.1.0
certifi==2023.7.22
charset-normalizer==3.3.2
click==8.1.7
dnspython==2.4.2
email-validator==2.1.0.post1
exceptiongroup==1.1.3
fastapi==0.104.1
frozenlist==1.4.0
gunicorn==21.2.0
h11==0.14.0
httpcore==0.18.0
httptools==0.6.1
httpx==0.25.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
multidict==6.0.4
openai==0.28.1
orjson==3.9.10
outcome==1.3.0.post0
packaging==23.2
pydantic==2.4.2
pydantic-extra-types==2.1.0
pydantic-settings==2.0.3
pydantic_core==2.10.1
PySocks==1.7.1
python-dotenv==1.0.0
python-multipart==0.0.6
PyYAML==6.0.1
requests==2.31.0
schedule==1.2.1
selenium==4.15.0
sniffio==1.3.0
sortedcontainers==2.4.0
starlette==0.27.0
tqdm==4.66.1
trio==0.22.2
trio-websocket==0.11.1
typing_extensions==4.8.0
ujson==5.8.0
urllib3==2.0.7
uvicorn==0.23.2
uvloop==0.19.0
watchfiles==0.21.0
websockets==12.0
wsproto==1.2.0
yarl==1.9.2
